Move the idl macro impl in a sub-module for better code organization
Changes:
  - separate tlv and idl macros impl
  - started adding unit tests for tlv part
