+++ rs-matter-macros-impl/Cargo.toml
@@ -8,9 +8,9 @@ edition = "2021"
 [dependencies]
 quote = "1.0.33"
 rs-matter-data-model={path = "../rs-matter-data-model", features = ["idl"]}
+syn = {version="2.0.48", features = ["extra-traits", "parsing", "proc-macro"]}
 proc-macro2 = "1.0.70"
 convert_case = "0.6.0"
 
 [dev-dependencies]
 assert-tokenstreams-eq = "0.1.0"
-syn = {version="2.0.48", features = ["extra-traits"]}
